requirements.txt
playwright==1.40.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.1.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy>=1.26.0
scikit-learn>=1.3.2
streamlit>=1.28.0
nest-asyncio>=1.5.8 